changes on git ignore

+++ app_backend/src/ocr/ocr.controller.ts
@@ -1,15 +1,15 @@
-import { Controller, Post, Body, UploadedFile, UseInterceptors, Get } from '@nestjs/common';
+import { Controller, Post, Body, UploadedFile, UseInterceptors, Get, BadRequestException } from '@nestjs/common';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { diskStorage } from 'multer';
 import { extname } from 'path';
 import { OcrService } from './ocr.service'; 
 import { LabTestEntity } from './labtest.entity';
 import { ExtractionEntity } from './extraction.entity';
+import * as fs from 'fs';
 
 @Controller('ocr')
 export class OcrController {
   constructor(private readonly ocrService: OcrService) {}
-
   @Post('process')
   @UseInterceptors(
     FileInterceptor('image', {
@@ -27,24 +27,31 @@ export class OcrController {
   async processImage(
     @UploadedFile() file: Express.Multer.File,
     @Body('id_patient') id: number,
-  )
-  : Promise<{ labtest: LabTestEntity; extractions: ExtractionEntity[] }>
-   {
-    const imagePath = `./src/ocr/img/${file.filename}`;
-
-    const { extractions, labtest } = await this.ocrService.processImage(imagePath, id);
-
-    const fs = require('fs');
-    fs.unlink(imagePath, (err) => {
-      if (err) {
-        console.error('Failed to delete image:', err);
-      } else {
-        console.log('Image deleted successfully:', imagePath);
+  ): Promise<{ labtest: LabTestEntity; extractions: ExtractionEntity[] }> {
+    try {
+      if (!file) {
+        throw new BadRequestException('No file uploaded');
       }
-    });
+      console.log('hi tehre');
+      const imagePath = `./src/ocr/img/${file.filename}`;
+      console.log(imagePath);
+      const { extractions, labtest } = await this.ocrService.processImage(imagePath, id);
 
-    return { labtest, extractions };
+      fs.unlink(imagePath, (err) => {
+        if (err) {
+          console.error('Failed to delete image:', err);
+        } else {
+          console.log('Image deleted successfully:', imagePath);
+        }
+      });
+
+      return { labtest, extractions };
+    } catch (error) {
+      console.error('Error processing image:', error);
+      throw new Error('Image processing failed');
+    }
   }
+
 
   @Post('save')
   async saveResults(
